typescript stuff
