+++ dbapi/.nvmrc
@@ -1,1 +1,1 @@
-6.9.0
+7.2.0
