+++ dbapi/bin/.gitignore
@@ -1,0 +1,4 @@
+# ignore everything in folder
+*
+*/
+!.gitignore
