+++ dbapi/package.json
@@ -1,12 +1,21 @@
 {
   "name": "dbapi",
   "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {},
-  "author": "",
+  "description": "Coliselia database api to allow interaction between SIG-Game components and common database",
+  "main": "bin/server.js",
+  "scripts": {
+    "start": "node bin/server.js",
+    "lint": "tslint --fix src/**/*.ts",
+    "compile": "tsc"
+  },
+  "author": "siggame",
   "license": "ISC",
   "dependencies": {
+    "@types/body-parser": "0.0.33",
+    "@types/config": "0.0.31",
+    "@types/express": "^4.0.34",
+    "@types/knex": "0.0.36",
+    "@types/winston": "0.0.32",
     "body-parser": "^1.15.2",
     "config": "^1.24.0",
     "dotenv": "^2.0.0",
@@ -15,5 +24,9 @@
     "knex": "^0.12.6",
     "pg": "^6.1.0",
     "winston": "^2.3.0"
+  },
+  "devDependencies": {
+    "tslint": "^4.0.2",
+    "typescript": "^2.2.0-dev.20161206"
   }
 }
