+++ dbapi/src/game_stats_api.ts
@@ -1,0 +1,9 @@
+import * as express from "express";
+
+let router = express.Router();
+
+router.get("/", (req, res) => {
+    res.status(400).send("UNIMPLEMENTED");
+});
+
+export { router as GameStatsApiRouter };
