+++ dbapi/src/match_api.ts
@@ -1,0 +1,13 @@
+import * as express from "express";
+
+let router = express.Router();
+
+router.get("/:id", (req, res) => {
+    res.status(400).send("UNIMPLEMENTED");
+});
+
+router.post("/api/v2/match/", function(req, res){
+    res.status(400).send("UNIMPLEMENTED");
+});
+
+export { router as MatchApiRouter };
