+++ dbapi/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+        "module": "es6",
+        "target": "es6",
+        "outDir": "bin/"
+    },
+    "include": [
+        "src/**/*"
+    ]
+}
