+++ dbapi/tslint.json
@@ -1,0 +1,11 @@
+{
+    "extends": "tslint:latest",
+    "rules": {
+        "curly": true
+    },
+    "jsRules": {
+        "curly": true
+    },
+    "rulesDirectory": [
+    ]
+}
