+++ docker-compose.yml
@@ -15,7 +15,7 @@ services:
       DB_USER: "coliselia"
       DB_PASS: "siggame"
       DB_NAME: "coliselia"
-      DB_HOST: "coliselia_db_1"
+      DB_HOST: "db"
       DB_PORT: "5432"
     build: dbapi/
     ports:
